Moved loading prompt.js into prompt.html
Having this in index.js resulted in the script loading before the page, but it
has to load after the page to access its elements. Having to remember that is
probably a bad thing for index.js anyway since prompt.html might later need it
loaded at a different place, or need extra scripts, and index.js isn't exactly
the kind of place you expect to need to put includes for someone else.