Added run.sh
Eventually I'll get around to setting up an actual profile for Firefox to use
so it doesn't constantly nag me about its fancy new reader mode and data
reporting policy. When that happens, having a run script that handles all of
that for me so I don't have to constantly type it in will be nice. Not having
to constantly scroll through the command history will already be nice, too.